Set default login mode to first authentication-mode option (#4921)
* Set default login mode to first authentication-mode option

authentication-mode options specified in backend, i.e. `token` (default),
`basic` or both, are retrieved from frontend, and login mode selections
are set in order of the options specified.

But `Kubeconfig` is set as default in frontend for now.

This commit enables to set default selection by operator's specification
of authentication-mode.

* Simplify logic

* Fix test

Co-authored-by: Sebastian Florek <s.florek91@gmail.com>