Adding override for ingress paths (#5276)
* Adding override for ingress paths

This PR is adding the possibility to override the default Ingress paths
with a new block of paths defined as YAML data structure. This allows to
configure the Ingress resource to work with the AWS ALB Ingress
Controller to automatically [redirect HTTP traffic to
HTTPS](https://kubernetes-sigs.github.io/aws-alb-ingress-controller/guide/tasks/ssl_redirect/)
on the load balancer.

* Implementing comments from desaintmartin